Player movement
A Player Movement sculpt was created for the capsule and basic left, right and jump controls were added, making sure the jump distance and move speed was editable.

Created a GroundCheck mechanic were all walkable surfaces were given a "Walkable" layer in order to create both walkable surfaces and purely decorational surfaces.

Some minor changes were added in order to make the movement smoother, for example creating a Physics Material 2D and setting the Friction to 0 and adding it to the Player's RigidBody component. This was implemented as when the player was jumping, he kept getting stuck like velcro to surfaces and with this easy fix, he slides off.

A CameraFollow script was created in order to make the camera follow the player, but I thought this was too choppy so I added an offset and used SmoothDamp to make the camera move more smoothly.